Benchmark 64 bit keys

+++ benchmarks/ShuffleBench.cu
@@ -53,28 +53,28 @@ static void sortArgsGenerator( benchmark::internal::Benchmark* b )
 }
 
 
-BENCHMARK_TEMPLATE( benchmarkFunction, FeistelBijectiveShuffle<thrust::device_vector<uint8_t>> )->Apply( argsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, FeistelBijectiveSortShuffle<thrust::device_vector<uint8_t>> )->Apply( sortArgsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, FeistelBijectiveScanShuffle<thrust::device_vector<uint8_t>> )->Apply( sortArgsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, SPNetworkBijectiveShuffle<thrust::device_vector<uint8_t>> )->Apply( argsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, SPNetworkBijectiveSortShuffle<thrust::device_vector<uint8_t>> )
+BENCHMARK_TEMPLATE( benchmarkFunction, FeistelBijectiveShuffle<thrust::device_vector<uint64_t>> )->Apply( argsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, FeistelBijectiveSortShuffle<thrust::device_vector<uint64_t>> )->Apply( sortArgsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, FeistelBijectiveScanShuffle<thrust::device_vector<uint64_t>> )->Apply( sortArgsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, SPNetworkBijectiveShuffle<thrust::device_vector<uint64_t>> )->Apply( argsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, SPNetworkBijectiveSortShuffle<thrust::device_vector<uint64_t>> )
     ->Apply( sortArgsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, SPNetworkBijectiveScanShuffle<thrust::device_vector<uint8_t>> )
+BENCHMARK_TEMPLATE( benchmarkFunction, SPNetworkBijectiveScanShuffle<thrust::device_vector<uint64_t>> )
     ->Apply( sortArgsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, LCGBijectiveShuffle<thrust::device_vector<uint8_t>> )->Apply( argsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, LCGBijectiveSortShuffle<thrust::device_vector<uint8_t>> )->Apply( sortArgsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, LCGBijectiveScanShuffle<thrust::device_vector<uint8_t>> )->Apply( sortArgsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, NoOpBijectiveShuffle<thrust::device_vector<uint8_t>> )->Apply( argsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, NoOpBijectiveSortShuffle<thrust::device_vector<uint8_t>> )->Apply( sortArgsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, NoOpBijectiveScanShuffle<thrust::device_vector<uint8_t>> )->Apply( sortArgsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, SortShuffle<thrust::device_vector<uint8_t>> )->Apply( sortArgsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, StdShuffle<std::vector<uint8_t>> )->Apply( argsGenerator );
-BENCHMARK_TEMPLATE( benchmarkFunction, FisherYatesShuffle<std::vector<uint8_t>> )->Apply( argsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, LCGBijectiveShuffle<thrust::device_vector<uint64_t>> )->Apply( argsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, LCGBijectiveSortShuffle<thrust::device_vector<uint64_t>> )->Apply( sortArgsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, LCGBijectiveScanShuffle<thrust::device_vector<uint64_t>> )->Apply( sortArgsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, NoOpBijectiveShuffle<thrust::device_vector<uint64_t>> )->Apply( argsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, NoOpBijectiveSortShuffle<thrust::device_vector<uint64_t>> )->Apply( sortArgsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, NoOpBijectiveScanShuffle<thrust::device_vector<uint64_t>> )->Apply( sortArgsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, SortShuffle<thrust::device_vector<uint64_t>> )->Apply( sortArgsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, StdShuffle<std::vector<uint64_t>> )->Apply( argsGenerator );
+BENCHMARK_TEMPLATE( benchmarkFunction, FisherYatesShuffle<std::vector<uint64_t>> )->Apply( argsGenerator );
 // Too slow
-// BENCHMARK_TEMPLATE( benchmarkFunction, LubyRackoffBijectiveShuffle<thrust::device_vector<uint8_t>> )->Apply( argsGenerator );
-// BENCHMARK_TEMPLATE( benchmarkFunction, LubyRackoffBijectiveSortShuffle<thrust::device_vector<uint8_t>> )->Apply( sortArgsGenerator );
-// BENCHMARK_TEMPLATE( benchmarkFunction, LubyRackoffBijectiveScanShuffle<thrust::device_vector<uint8_t>> )->Apply( sortArgsGenerator );
-// BENCHMARK_TEMPLATE(benchmarkFunction, GPUSwapShuffle<uint8_t>)->Apply(argsGenerator);
+// BENCHMARK_TEMPLATE( benchmarkFunction, LubyRackoffBijectiveShuffle<thrust::device_vector<uint64_t>> )->Apply( argsGenerator );
+// BENCHMARK_TEMPLATE( benchmarkFunction, LubyRackoffBijectiveSortShuffle<thrust::device_vector<uint64_t>> )->Apply( sortArgsGenerator );
+// BENCHMARK_TEMPLATE( benchmarkFunction, LubyRackoffBijectiveScanShuffle<thrust::device_vector<uint64_t>> )->Apply( sortArgsGenerator );
+// BENCHMARK_TEMPLATE(benchmarkFunction, GPUSwapShuffle<uint64_t>)->Apply(argsGenerator);
 
 
 BENCHMARK_MAIN();
